+++ src/main/java/video/infrastructure/datasource/data/RentalRecord.java
@@ -1,5 +1,5 @@
 package video.infrastructure.datasource.data;
 
-public record RentalRecord(String customerName, double totalAmount) {
+public record RentalRecord(String customerName, int totalAmount) {
 
 }
